Fold agentConfigBasic into useAgentBuilderForm
The draft types, empty draft, and completeness check belong to the form.
isBasicConfigComplete becomes per-section validators (draftValid,
machineSourceValid, mcpServerValid, mcpServerNamesUnique) private to the
hook; other modules import the draft types from useAgentBuilderForm.
