+++ frontend/apps/web/src/components/agents/useAgentBuilderForm.ts
@@ -1,15 +1,69 @@
+import type { ToolPermissionSelection } from '@omnara/sdk'
 import { useEffect, useState } from 'react'
 
-import {
-  type BasicConfig,
-  type BasicMachineSource,
-  type BasicMcpServer,
-  emptyBasicConfig,
-  isBasicConfigComplete,
-} from '@/components/agents/agentConfigBasic'
 import type { BasicConfigSession } from '@/components/agents/agentConfigBasicYaml'
 import type { ModelSelection } from '@/components/agents/AgentConfigModelField'
 import type { BasicTool } from '@/components/agents/AgentConfigToolsField'
+import {
+  type EnvOverlayRow,
+  envOverlayRowsValid,
+  optionalPositiveInt32Valid,
+  type ProviderOptionsDraft,
+  type SecretEnvOverlayRow,
+  secretEnvOverlayRowsValid,
+} from '@/components/machines/machineOverrides'
+
+export type McpAuthType = 'none' | 'oauth' | 'bearer' | 'sigv4'
+
+export interface BasicMcpServer {
+  id: string
+  name: string
+  url: string
+  permission: ToolPermissionSelection
+  defaultEnabled: boolean
+  authType: McpAuthType
+  secretId: string
+  service: string
+  region: string
+}
+
+export type MachineSourceKind = 'pool' | 'machine'
+
+export interface BasicMachineSource {
+  id: string
+  kind: MachineSourceKind
+  name: string
+  provider: string
+  managementKind: string
+  defaultCwd: string
+  initialNumMachines: string
+  maxMachines: string
+  machineCpu: string
+  machineMemoryMb: string
+  providerOptions: ProviderOptionsDraft
+  envRows: EnvOverlayRow[]
+  secretEnvRows: SecretEnvOverlayRow[]
+}
+
+export interface BasicConfig {
+  instruction: string
+  providerConfig: string
+  modelName: string
+  machineSources: BasicMachineSource[]
+  tools: BasicTool[]
+  mcpServers: BasicMcpServer[]
+  skillIds: string[]
+}
+
+const emptyBasicConfig: BasicConfig = {
+  instruction: '',
+  providerConfig: '',
+  modelName: '',
+  machineSources: [],
+  tools: [],
+  mcpServers: [],
+  skillIds: [],
+}
 
 export function useAgentBuilderForm(
   session: BasicConfigSession,
@@ -24,7 +78,7 @@ export function useAgentBuilderForm(
     unavailableSkillIds.length > 0 ||
     unavailableSourceIds.length > 0 ||
     modelUnavailable ||
-    !isBasicConfigComplete(draft)
+    !draftValid(draft)
 
   useEffect(() => {
     onYamlChange(session.apply(draft), blocked)
@@ -63,4 +117,52 @@ export function useAgentBuilderForm(
     reportUnavailableSourceIds: setUnavailableSourceIds,
     reportUnavailableSkillIds: setUnavailableSkillIds,
   }
+}
+
+function draftValid(draft: BasicConfig) {
+  return (
+    draft.instruction.trim() !== '' &&
+    draft.providerConfig.trim() !== '' &&
+    draft.modelName.trim() !== '' &&
+    draft.machineSources.every(machineSourceValid) &&
+    mcpServerNamesUnique(draft.mcpServers) &&
+    draft.mcpServers.every(mcpServerValid)
+  )
+}
+
+const positiveIntegerPattern = /^[1-9][0-9]*$/
+
+function machineCountValid(value: string) {
+  return value === '' || positiveIntegerPattern.test(value)
+}
+
+function machineSourceValid(source: BasicMachineSource) {
+  return (
+    source.name.trim() !== '' &&
+    envOverlayRowsValid(source.envRows) &&
+    secretEnvOverlayRowsValid(source.secretEnvRows) &&
+    (source.kind === 'machine' ||
+      (machineCountValid(source.initialNumMachines) &&
+        machineCountValid(source.maxMachines) &&
+        optionalPositiveInt32Valid(source.machineCpu) &&
+        optionalPositiveInt32Valid(source.machineMemoryMb)))
+  )
+}
+
+const mcpServerNamePattern = /^[a-zA-Z][a-zA-Z0-9-]{0,31}$/
+
+function mcpServerValid(server: BasicMcpServer) {
+  return (
+    mcpServerNamePattern.test(server.name.trim()) &&
+    server.url.trim() !== '' &&
+    (server.authType === 'none' ||
+      (server.secretId.trim() !== '' &&
+        (server.authType !== 'sigv4' ||
+          (server.service.trim() !== '' && server.region.trim() !== ''))))
+  )
+}
+
+function mcpServerNamesUnique(servers: BasicMcpServer[]) {
+  const names = servers.map((server) => server.name.trim())
+  return new Set(names).size === names.length
 }
